add AutoConfiguration for mock connection property
which can be enabled by property
