+++ tests/samples-messaging-amqp/src/test/resources/application.yml
@@ -1,0 +1,3 @@
+verifier:
+  amqp:
+    mockConnection: true
