+++ tests/spring-cloud-contract-stub-runner-amqp/src/test/resources/application.yml
@@ -1,2 +1,10 @@
+// tag::stub_runner_config[]
 stubrunner.repositoryRoot: classpath:m2repo/repository/
 stubrunner.ids: org.springframework.cloud.contract.verifier.stubs.amqp:spring-cloud-contract-amqp-test:0.4.0-SNAPSHOT:stubs
+// end::stub_runner_config[]
+
+// tag::rabbit_mock_connection_config[]
+verifier:
+  amqp:
+    mockConnection: true
+// end::rabbit_mock_connection_config[]
